requirements.txt
django>=5.0,<6.0
python-telegram-bot>=22.0,<23.0
python-dotenv>=1.0.1,<2.0.0
